drop support for -logfile option
this option was useless: you could just as well redirect stdout to a
file in your shell with exactly the same effect.

now, because we want to log issues with config file loading (see issue
commandline flags are parsed. it would then be difficult to
*reconfigure* the lumberjack logger when the -logfile flag is parsed,
so we just don't and remove the option.

we also split the logging function out of main so it's
cleaner. eventually, we could call it multiple times to support the
-logfile flag again, but this was deemed unnecessary.