start using godep
Trying to make sure we have a way to pin dependencies during
development. This doesn't fix fundamental issues with go get: that's a
broader problem with the go ecosystem that dep can't fix directly. We
fix this with CI right now, it's a packaging problem.

We also don't commit vendor/ as I object to the duplication of code. I
assume that the presence of the .lock file provides sufficient
protection against spurious upstream changes (e.g. with -vendor-only)
that we don't need to keep a local copy.