document research on database format
I have tried reverse-engineering the Kobo database format, and it's a
real mess. The `content` table, for example, has about 90 *columns*,
most of which are basically useless. There are multiple rows per book,
one per book but also one per chapter per book (instead of having,
say, a separate `chapters`) table.

There is a lot of user tracking junk in there as well, which tells me
the reader is spying on its users and reporting that information back
to headquarters, something which we already knew about, but it's kind
of creepy to see in action.

Even worse: those results are completely useless in trying to improve
the current update mechanisms. Changes to the database are not picked
up by the GUI at all, and changes in the GUI are not written in the
database until a reboot or USB connect. This will make it hard, if not
impossible to find a better way to update the contents than the
current `fake-usb-connect` mechanism. Collections/shelves (#20) will
also be difficult to implement, unless the DB is updated *while the
USB dialog is open*, which is really not easy to do, because the
database then gets unmounted!

So this was a full day of work wasted, thanks to Kobo's failed
engineering. It's a miracle this thing works at all, when you think
about it. This makes [okreader] really interesting again...

[okreader]: https://github.com/lgeek/okreader/