try to workaround gitlab.com artifact expiration policy
According to [the gitlab-ci.yml][1] reference, artifacts expire after
30 days by default. By setting this to `never`, we have a better
chance at not expiring those artifacts.

 [1]: https://docs.gitlab.com/ee/ci/yaml/README.html#artifactsexpire_in

Normally, latest artifacts should be protected against deletion like
this, but that feature is disabled on GitLab.com.

Worst case, we can also schedule a pipeline to run every month to
refresh those artifacts, but that would be somewhat against the spirit
of resource limits on gitlab.com.

Obviously, this commit should trigger a build which will resolve the
issue in the short term, but hopefully this will also be fixed in the
long term now.

Closes: #38