make a section about related projects
There is more than koreader that runs on Kobos, so it might be good to
introduce those as well. But start by rephrasing the comment from
@AlMartin80: instead of linking to the mobileread.com forum, we just
link to the KOreader wiki and use correct Markdown markup instead of
bare URLs.

The downside is we move this all the way to the end, but the Wallabako
README's first job is not to promote other projects but to document
Wallabako itself.

See: #35