replace "kobo" references by "nickel" where appropriate
We prefer to use "nickel" internally because a lot of things actually
without "Nickel" (the default Kobo UI) on the Kobo. This is liberating
and shows how that's just one another (and kind of buggy)
implementation of a book reader UI that we don't necessarily have to
deal with all the time...

In fact, this makes me realize I should make wallabako more of
device-agnostic program. For example, it could be used with koreader
wherever *that* program is ported, which is a lot of devices, way more
than just Kobos.