port to wallabago v6
This required a patch because we stumbled upon 'null' errors as
described in:

https://github.com/Strubbl/wallabago/pull/19

Also, this was hell. It took me a while just to figure out I had to
"replace" the module, but then that wasn't enough. Because I wasn't
using that /v6 hack everywhere, I would get this weird error message
on replace:

    replace github.com/simonfrey/wallabago: version "v1.0.1-0.20210117162249-afac782761b4" invalid: go.mod has post-v1 module path "github.com/Strubbl/wallabago/v6" at revision afac782761b4

That was totally opaque and I couldn't figure out how to fix it. Turns
out I had to specify the /v6 in the import *and* go mod.

This also required handling the error from the GetAuthTokenHeader()
function, but that's really minor in comparison.