switch to pure go sqlite implementation
This pulls in a metric ton of dependencies, unfortunately. It looks
like "pure go" here means "we transpiled sqlite to go", which is not
exactly a win for us.

The alternative (sqlittle) has a very different API, so this is the
"easy" fix for #43.

The cgo implementation is kept behind a feature flag (`--tags
sqlite3`) which can be used if we want to restore the previous
behavior.