make cat logger work
The trick was that commands don't inherit stdout/stderr in go, so
it *was* working, I was just not seeing it.

To be fair with ChatGPT, it *did* have that trick in there, just
completely unexplained (and actually not relevant for my final use
case, just coincidentally allowing me to confirm this works here).