use first row as a fbink status space, mark start/end more clearly
The point of this is to make the "does fbink work" check actually do
something useful. This is to workaround the fact I don't really have a
clue how to check it actually works yet, but I also think it's kind of
useful to show wtf is going on with a small-ish heading.