fbink: fix path in wallabako's caller
The implementation of on-screen display would frequently fail to
run correctly, particularly when it would be started from udev. It
seems that, from there, the PATH environment is just completely empty,
which, naturally, makes it... hard for wallabako to find the fbink
binary.

It seems like my original implementation in 7974548 (implement basic
fbink output (#49), 2023-06-09) already had this hack to try to add to
PATH, but that didn't work: I suspect exec.Command doesn't look at the
provided Env for the PATH environment. This is a bit counter-intuitive
to me, but actually makes sense: Env is passed *to* the subprocess,
but the PATH used is the one from the *calling* process.

(I haven't actually confirmed this, but I can definitely confirm that
fixing this in the script udev calls actually works. A golang-only fix
would be to change wallabako's PATH directly, but that seems more
intrusive and unnecessary for sane platforms where PATH is properly
set.)