fix fbink error reporting
Previously, we'd get `nil` as an error message when fbink would fail
to start, which would be extremely confusing. That's because we were
using the wrong `err` variable. Just get rid of that too generic `err`
and properly name error variables that are used elsewhere than just
the next line over.