fix -config parameter
It seems our -config flag has been broken for years. Not only someone
noticed, but they filed an issue and I said "bah, no, that thing
works".

Then today, I'm looking again at the code, particularly around the
fbink stuff, and noticed it just doesn't make any sense, as the
command line flags are parsed *after* the config file is parsed.

Even worse, the config parameter is just actually never used, as it's
clobbered by the configuration file parser anyway. It's why we were
using a pointer everywhere, which should have been a red flag anyway.

So now, instead, we parse the command line twice. I'm not even sure
that works, but it sure does seem like it. I tested that by using
`-debug` on a commandline with a `-config` file that had `Debug:
false` in it, and was still getting debugging messages after. So, yay?

Closes: #29