handle interrupts in download routine
When we download multiple files, instead of just completely stopping
in the middle of a download when receiving a signal, gracefully wait
for the current downloads to finish instead.

We only handle SIGHUP ("terminal gone"), SIGTERM ("kill") and
SIGINT ("control-c") here, and leave SIGQUIT ("crash and burn") alone,
the reasoning for the latter is that SIGQUIT is deliberately designed
to crash the program, and we want to keep that.

This doesn't actually interrupt in-progress downloads, that said,
which may cause stickiness.

Partly (but likely not fully) addresses #33.