remove files on I/O errors
I have briefly considered just doing this on timeout, but on second
thought, I don't see why. If we failed to download the file, we
failed, and whatever is left should be trashed.

In fact, the *only* reason to keep the file would be precisely if we
had a timeout. In that case, we might actually have good stuff on
disk and, if we supported resuming downloads (which we don't), we
could resume from there.

According to GPT-4o, this is how I would detect a timeout:

    if netErr, ok := err.(net.Error); ok && netErr.Timeout() { ... }

This should hopefully fix most cases where we create empty files.

See: #33